[GH-161] Fixed cast error of Long in debezium time converter

+++ kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/convert/logicaltype/DebeziumLogicalConverters.java
@@ -172,7 +172,7 @@ public class DebeziumLogicalConverters {
 
     @Override
     public String convert(Object kafkaConnectObject) {
-      java.util.Date date = new java.util.Date((Long) kafkaConnectObject);
+      java.util.Date date = new java.util.Date((Integer) kafkaConnectObject);
       return getBqTimestampFormat().format(date);
     }
   }
